Reorganise files to use folder structure
