+++ server/views/pages/visitors/visitors.njk
@@ -1,5 +1,5 @@
 {% set activeLink = 'visitors-link' %}
-{% extends "../partials/layout.njk" %}
+{% extends "partials/layout.njk" %}
 {% from "govuk/components/table/macro.njk" import govukTable %}
